+++ kuma-cli/src/cli.rs
@@ -26,6 +26,18 @@ pub(crate) struct Cli {
     #[arg(long, global = true)]
     pub mfa_token: Option<String>,
 
+    /// The MFA secret. Used to generate a tokens for logging into Uptime Kuma (alternative to a single use mfa_token).
+    #[arg(long, global = true)]
+    pub mfa_secret: Option<String>,
+
+    /// Log in using an jwt auth token (alternative to using username and password, does not require a mfa token). Can be obtained using the `login` command.
+    #[arg(long, global = true)]
+    pub auth_token: Option<String>,
+
+    /// Store the auth token after a successful login. The token will be used for subseqent logins bypassing the need for a mfa token.
+    #[arg(long = "store-token", default_value_t = false, global = true)]
+    pub store_auth_token: bool,
+
     /// Add a HTTP header when connecting to Uptime Kuma.
     #[arg(long = "header", value_name = "KEY=VALUE", global = true)]
     pub headers: Vec<String>,
@@ -80,6 +92,8 @@ impl From<Cli> for Config {
             .set_override_option("username", value.username.clone()).unwrap()
             .set_override_option("password", value.password.clone()).unwrap()
             .set_override_option("mfa_token", value.mfa_token.clone()).unwrap()
+            .set_override_option("mfa_secret", value.mfa_secret.clone()).unwrap()
+            .set_override_option("auth_token", value.auth_token.clone()).unwrap()
             .set_override_option(
                 "headers",
                 match value.headers.is_empty() {
@@ -132,5 +146,10 @@ pub(crate) enum Commands {
     DockerHost {
         #[command(subcommand)]
         command: Option<crate::docker_host::Command>,
+    },
+    /// Authenticate with the uptime kuma server
+    Login {
+        #[command(flatten)]
+        command: crate::login::Command,
     },
 }