+++ kuma-cli/src/main.rs
@@ -5,6 +5,7 @@ use kuma_client::Config;
 
 mod cli;
 mod docker_host;
+mod login;
 mod maintenance;
 mod monitor;
 mod notification;
@@ -33,6 +34,7 @@ async fn main() {
         }
         Some(Commands::StatusPage { command }) => status_page::handle(command, &config, &cli).await,
         Some(Commands::DockerHost { command }) => docker_host::handle(command, &config, &cli).await,
+        Some(Commands::Login { command }) => login::handle(command, &config, &cli).await,
         None if cli.shadow => kuma_client::build::print_build_in(),
         None => {}
     };